refactor: Extract ShowReconciler interface to api module
* Define `ShowReconciler` interface and associated models in the `api` module to decouple contract from implementation

* feat: Move `ShowReconciler` interface, `ShowResolveOutcome`, and `ReconciliationResult` to `data:shows:api`
* refactor: Rename implementation to `DefaultShowReconciler` and apply `@ContributesBinding` for dependency injection
* refactor: Update `ShowResolveOutcome.Resolved` to hold a `tmdbId` (Long) instead of an internal database ID
* refactor: Remove `showIdResolver` dependency and query the database directly for existing show mappings
* chore: Add `data:accountManager:api` as a dependency to `data:shows:api` for `AccountProvider` access
* test: Update `ShowReconcilerTest` to align with the renamed class and verify TMDB ID resolution logic
