+++ data/shows/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/shows/implementation/DefaultShowReconciler.kt
@@ -5,36 +5,31 @@ import com.thomaskioko.tvmaniac.core.logger.Logger
 import com.thomaskioko.tvmaniac.core.networkutil.api.model.ApiResponse
 import com.thomaskioko.tvmaniac.db.Id
 import com.thomaskioko.tvmaniac.db.Provider
-import com.thomaskioko.tvmaniac.db.ShowId
-import com.thomaskioko.tvmaniac.db.ShowIdResolver
 import com.thomaskioko.tvmaniac.db.TmdbId
 import com.thomaskioko.tvmaniac.db.TvManiacDatabase
+import com.thomaskioko.tvmaniac.shows.api.ReconciliationResult
+import com.thomaskioko.tvmaniac.shows.api.ShowReconciler
+import com.thomaskioko.tvmaniac.shows.api.ShowResolveOutcome
 import com.thomaskioko.tvmaniac.tmdb.api.TmdbShowsNetworkDataSource
 import dev.zacsweers.metro.AppScope
-import dev.zacsweers.metro.Inject
+import dev.zacsweers.metro.ContributesBinding
 import dev.zacsweers.metro.SingleIn
 
-public sealed class ShowResolveOutcome {
-    public data class Resolved(public val showId: Id<ShowId>) : ShowResolveOutcome()
-    public data object Skipped : ShowResolveOutcome()
-}
-
-@Inject
 @SingleIn(AppScope::class)
-public class ShowReconciler(
+@ContributesBinding(AppScope::class)
+public class DefaultShowReconciler(
     private val tmdbDataSource: TmdbShowsNetworkDataSource,
     private val database: TvManiacDatabase,
-    private val showIdResolver: ShowIdResolver,
     private val logger: Logger,
-) {
+) : ShowReconciler {
 
-    public suspend fun reconcile(
+    override suspend fun reconcile(
         tmdbId: Long?,
         imdbId: String?,
         title: String?,
         providerShowId: String?,
         provider: AccountProvider,
-        result: ReconciliationResult = ReconciliationResult(matched = 0, tmdbMissing = 0),
+        result: ReconciliationResult,
     ): Pair<ShowResolveOutcome, ReconciliationResult> {
         val resolvedTmdbId = when {
             tmdbId != null -> tmdbId
@@ -49,10 +44,10 @@ public class ShowReconciler(
             provider = provider,
         )
 
-        val showId = showIdResolver.showIdForTmdbId(resolvedTmdbId)
+        database.tvShowQueries.getShowIdByTmdbId(Id<TmdbId>(resolvedTmdbId)).executeAsOneOrNull()
             ?: return ShowResolveOutcome.Skipped to result.copy(tmdbMissing = result.tmdbMissing + 1)
 
-        return ShowResolveOutcome.Resolved(showId) to result.copy(matched = result.matched + 1)
+        return ShowResolveOutcome.Resolved(resolvedTmdbId) to result.copy(matched = result.matched + 1)
     }
 
     private suspend fun resolveViaImdb(imdbId: String): Long? =
