--- data/shows/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/shows/implementation/ReconciliationResult.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.shows.implementation
-
-public data class ReconciliationResult(
-    public val matched: Int,
-    public val tmdbMissing: Int,
-)
